Rename the backend packages: 'handler'->'handlers', 'util'->'utils'
