+++ backend/handlers/handlers.go
@@ -1,4 +1,4 @@
-package handler
+package handlers
 
 import (
 	"github.com/Slowers-team/Slowers-App/database"
